+++ api/src/main/java/de/maxhenkel/voicechat/api/events/ClientSoundEvent.java
@@ -8,10 +8,22 @@ public interface ClientSoundEvent extends ClientEvent {
     /**
      * The unencoded audio data.
      * <br/>
-     * Returns an empty array in case of the end of transmission
+     * <s>Returns an empty array in case of the end of transmission</s>
      *
      * @return the raw 16 bit PCM audio frame
      */
     short[] getRawAudio();
+
+    /**
+     * Overrides the actual audio data that's sent to the server.
+     *
+     * @param rawAudio the raw 16 bit PCM audio frame
+     */
+    void setRawAudio(short[] rawAudio);
+
+    /**
+     * @return if the player is whispering
+     */
+    boolean isWhispering();
 
 }