+++ fabric/changelog.md
@@ -1,1 +1,1 @@
-- Added force voicechat config option
+- Added client audio data override to API